+++ example/regular_callback.dart
@@ -3,12 +3,11 @@
 
 import 'base/lock_example_base.dart';
 
-
 /// This is the most common case in which the lock manager can be useful
 /// If you se the output, the calls with the values of 5, 7, 1 and 3 are
 /// all mixed up.
 main() {
-  WithoutLockExample withoutLockExample = new WithoutLockExample();
+  WithoutLockExample withoutLockExample = WithoutLockExample();
   withoutLockExample.foo(5).then((int value) => print(value));
   withoutLockExample.foo(7).then((int value) => print(value));
   withoutLockExample.foo(1).then((int value) => print(value));