+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.1.2
+
+- Fix on Waitlock, final solution. Breaking change.
+  The function to be called must be passed as waitLock parameter.
+
 ## 0.1.1
 
 - Fix on Waitlock. It was not working with dart2js. Temporary solution.